backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
pydantic-settings==2.1.0
requests==2.32.5
beautifulsoup4==4.14.3
yfinance==1.2.0
pandas==3.0.1
aiosqlite==0.19.0
apscheduler==3.10.4
websockets==16.0
